fixed some issues with registration and showing the map on login

+++ src/database/queries/createUserMapLink.sql
@@ -3,6 +3,6 @@ insert into kingmaker.person_map(
     map_id
 )
 values(
-    $1,
-    $2
+    $1::integer,
+    $2::integer
 )
