+++ src/database/queries/getMapLink.sql
@@ -1,3 +1,3 @@
 select map_link_code
 from kingmaker.map_links
-where map_id = $1
+where map_id = $1::integer
